+++ Torrent.cpp
@@ -84,7 +84,7 @@ void Torrent::setFileList(const valueDictionary& torrent)
 	if (info.count("name") && info.count("length"))
 	{
 		resObj.filePath = boost::get<std::string>(info.at("name"));
-		resObj.fileSize = static_cast<long long>(boost::get<integer>(info.at("length")));
+		resObj.fileSize = static_cast<long long>(boost::get<long long>(info.at("length")));
 		fileList.push_back(resObj);
 	}
 	//if torrent contains multiple files
@@ -103,7 +103,7 @@ void Torrent::setFileList(const valueDictionary& torrent)
 			//get file size and offset
 			if (fileData.count("length"))
 			{
-				long long size = static_cast<long>(boost::get<integer>(fileData.at("length")));
+				long long size = static_cast<long>(boost::get<long long>(fileData.at("length")));
 				resObj.fileSize = size;
 				resObj.setReadableFileSize();
 				resObj.fileOffset = runningOffset;
